master

+++ modules/rbac/controllers/RoleController.php
@@ -142,15 +142,14 @@ class RoleController extends \backend\components\Controller
 
         $model = $this->findModel($id);
 
-        $post = Yii::$app->request->post();
-        if (
-            $model->load($post)
-            && $model->save()
-        ) {
-            $model->deleteAllChildren();
-            $model->addChildren($post['child-roles-names']);
-            $model->addChildren($post['child-permissions-names']);
-            return $this->redirect(['view', 'id' => $model->id]);
+        if (Yii::$app->request->isPost) {
+            $post = Yii::$app->request->post();
+            if ($model->load($post) && $model->save()) {
+                $model->deleteAllChildren();
+                $model->addChildren($post['child-roles-names']);
+                $model->addChildren($post['child-permissions-names']);
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
         }
 
         $rolesNotOfThisRole = $rolesOfThisRole = $permissionsNotOfThisRole = $permissionsOfThisRole = [];
